% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{bucketWholesale}
\alias{bucketWholesale}
\title{bucketWholesale}
\usage{
bucketWholesale(position = position, t = t, EAD = EAD,
  LGDAddon = LGDAddon, LGD = LGD, PD = PD, rho = rho,
  equation = equation)
}
\arguments{
\item{position}{}

\item{t}{}

\item{EAD}{}

\item{LGDAddon}{}

\item{LGD}{}

\item{PD}{}

\item{rho}{}

\item{equation}{}
}
\description{
bucketWholesale
}
